+++ src/test/java/com/jcraft/jsch/KeyPairIT.java
@@ -24,7 +24,7 @@ public class KeyPairIT {
 
     @BeforeAll
     public static void beforeAll() {
-        JSch.setLogger(Slf4jLogger.getInstance());
+        JSch.setLogger(new Slf4jLogger());
     }
 
     @ParameterizedTest